tests: Move source annotation coverage
Batch display tests cover rendering after source annotation moved into its
own batch module.

That leaves annotation behavior validated from a display-named test file,
so the test tree no longer mirrors the module boundary used by runtime
code.

This commit moves the annotation-focused tests into test_source_annotation
and updates their monkeypatch targets to the source_annotation module.

A follow-up architecture guard will pin the boundary so display modules do
not regain batch-source annotation ownership.
